style: fix clippy warnings across the tree
Apply clippy's mechanical suggestions (collapsible match/if, manual
split_once, sort_by_key, is_multiple_of, needless clone/cast, etc.) and
guard the few intentional cases with targeted #[allow]s: the many-arg
config/draw entry points, parse_bool's unit error, and the recursion-only
parameter. Introduce SearchHit/SearchBatch type aliases for the search
channel, and switch two `.lines().flatten()` loops to `.map_while` so a
persistent read error can't spin forever. No behaviour change.

+++ src/tui/konami.rs
@@ -182,7 +182,7 @@ fn draw_matrix(f: &mut Frame, elapsed_ms: u64, done: bool, success: bool) {
         '0','1','ﾊ','ﾐ','ﾋ','ｰ','ｳ','ｼ','ﾅ','ﾓ','ﾆ','ｻ','ﾜ','ﾂ','ｵ','ﾘ','ｱ','ﾎ','ﾃ',
     ];
     for col in 0..w {
-        let speed = 40 + (hash(col as u64) % 80) as u64;
+        let speed = 40 + (hash(col as u64) % 80);
         let head_y = ((elapsed_ms / speed) % (h as u64 * 2)) as i32;
         let trail = 6 + (hash(col as u64 ^ 0xF00D) % 8) as i32;
         for d in 0..trail {
